+++ R/run_query.R
@@ -12,6 +12,25 @@
 #'   (If no database connection was used, note this may spin up a new one
 #'   dependending on your configuration. See \code{\link{last_connection}}.)
 run_query <- function(query, conn = last_connection()) {
+  # Attempts to connect with the database.  If it fails, it will reload the connection and try again.
+  tryCatch(
+    db_query(query, conn),
+    error = function(c) {
+      message("* Error Occurred, Reloading Connection...")
+      tryCatch({
+        suppressMessages(conn <- postgresql_connection(getOption('database.yml'), strict = FALSE));
+        db_query(query, conn);
+      },
+      error = function(c) {
+        message('* Reloading the Connection did not fix the Error.'); 
+        message(c) 
+      }
+      )
+    }
+  )
+}
+
+db_query <- function(query, conn = last_connection()) {
   if (missing(query)) return(last_connection())
   stopifnot(is(conn, 'JDBCConnection')) 
   stopifnot(is.character(query))